Fixed Item pickups and some extra tweaks
I've been looking for a keep alive as well but couldn't find anything.
After implementing the response to a RemoveItemOwner packet the server
no longer randomly disconnects us, this probably means Terraria has no
dedicated keep alives but a remote client hasn't sent anything to the
server in a while the server will treat is as disconnected.